+++ Project work/include/opencl.h
@@ -3,10 +3,10 @@
 
 char* readKernelSource(const char* filename);
 
-void cl_max_sort(unsigned long* arr, int size, int group);
+void cl_max_sort(int* arr, int size, int group);
 
-void cl_quick_sort(unsigned long* arr, int size);
+void cl_quick_sort(int* arr, int size);
 
-void cl_radix_sort(unsigned long* arr, int size, unsigned long max);
+void cl_radix_sort(int* arr, int size, int max);
 
 #endif
