+++ Project work/include/sorting.h
@@ -1,18 +1,18 @@
 #ifndef SORTING_H
 #define SORTING_H
 
-void swap(unsigned long* a, unsigned long* b);
+void swap(int* a, int* b);
 
-void max_sort(unsigned long* arr, int size);
+void max_sort(int* arr, int size);
 
-void quick_sort(unsigned long* arr, int low, int high);
+void quick_sort(int* arr, int low, int high);
 
-int partition(unsigned long* arr, int low, int high);
+int partition(int* arr, int low, int high);
 
-void radix_sort_quick(unsigned long* arr, int size, unsigned long max);
+void radix_sort_quick(int* arr, int size, int max);
 
-void quick_sort_by_digit(unsigned long* arr, int low, int high, unsigned long exp);
+void quick_sort_by_digit(int* arr, int low, int high, int exp);
 
-int partition_by_digit(unsigned long* arr, int low, int high, unsigned long exp); 
+int partition_by_digit(int* arr, int low, int high, int exp); 
 
 #endif
