+++ Project work/kernel/quick.cl
@@ -1,4 +1,4 @@
-__kernel void partition(__global unsigned long* arr, int low, int high, __global int* pivot_index, unsigned long pivot) 
+__kernel void partition(__global int* arr, int low, int high, __global int* pivot_index, int pivot) 
 {
     int i = low - 1;
     for (int j = low; j <= high - 1; j++) 
@@ -6,22 +6,22 @@ __kernel void partition(__global unsigned long* arr, int low, int high, __global
         if (arr[j] <= pivot) 
         {
             i++;
-            unsigned long temp = arr[i];
+            int temp = arr[i];
             arr[i] = arr[j];
             arr[j] = temp;
         }
     }
-    unsigned long temp = arr[i + 1];
+    int temp = arr[i + 1];
     arr[i + 1] = arr[high];
     arr[high] = temp;
     *pivot_index = i + 1;
 }
 
-__kernel void insertion_sort(__global unsigned long* arr, int low, int high) 
+__kernel void insertion_sort(__global int* arr, int low, int high) 
 {
     for (int i = low + 1; i <= high; i++) 
     {
-        unsigned long key = arr[i];
+        int key = arr[i];
         int j = i - 1;
         while (j >= low && arr[j] > key) 
         {
